Fix for ISSUE #3149: trying to change the Language
