+++ geonode/base/utils.py
@@ -29,14 +29,6 @@ from django.conf import settings
 
 # Geonode functionality
 from geonode.documents.models import ResourceBase
-from django.shortcuts import render_to_response
-from django.template import RequestContext
-from django.http import HttpResponseRedirect
-from django.core.exceptions import PermissionDenied
-from geonode.documents.models import Document
-from geonode.layers.models import Layer
-from geonode.maps.models import Map
-from .forms import BatchEditForm
 
 
 def delete_orphaned_thumbs():
@@ -54,42 +46,3 @@ def delete_orphaned_thumbs():
                 os.remove(fn)
             except OSError:
                 print 'Could not delete file %s' % fn
-
-
-def batch_modify(request, ids, model):
-    if not request.user.is_superuser:
-        raise PermissionDenied
-    if model == 'Document':
-        Resource = Document
-    if model == 'Layer':
-        Resource = Layer
-    if model == 'Map':
-        Resource = Map
-    template = 'base/batch_edit.html'
-    if request.method == 'POST':
-        form = BatchEditForm(request.POST)
-        if form.is_valid():
-            for resource in Resource.objects.filter(id__in=ids.split(',')):
-                resource.group = form.cleaned_data['group'] or resource.group
-                resource.owner = form.cleaned_data['owner'] or resource.owner
-                resource.category = form.cleaned_data['category'] or resource.category
-                resource.license = form.cleaned_data['license'] or resource.license
-                resource.date = form.cleaned_data['date'] or resource.date
-                resource.language = form.cleaned_data['language'] or resource.language
-                new_region = form.cleaned_data['regions']
-                if new_region:
-                    resource.regions.add(new_region)
-                resource.save()
-            return HttpResponseRedirect(
-                '/admin/{model}s/{model}/'.format(model=model.lower()))
-        return render_to_response(template, RequestContext(request, {
-            'form': form,
-            'ids': ids,
-            'model': model,
-        }))
-    form = BatchEditForm()
-    return render_to_response(template, RequestContext(request, {
-        'form': form,
-        'ids': ids,
-        'model': model,
-    }))
