+++ geonode/documents/tests.py
@@ -393,7 +393,7 @@ class DocumentsTest(TestCase):
         for resource in resources:
             self.assertEquals(resource.date, date)
         # test language change
-        language = 'Non-existing'
+        language = 'eng'
         response = self.client.post(
             reverse(view, args=(ids,)),
             data={'language': language},
