+++ geonode/layers/tests.py
@@ -825,7 +825,7 @@ class LayersTest(TestCase):
         for resource in resources:
             self.assertEquals(resource.date, date)
         # test language change
-        language = 'Non-existing'
+        language = 'eng'
         response = self.client.post(
             reverse(view, args=(ids,)),
             data={'language': language},
